# openapi.yaml
openapi: 3.0.3
info:
  title: LangGraph Multi-Agent System API
  description: |
    A comprehensive multi-agent workflow system built with FastAPI and LangGraph.

    This API provides endpoints for managing tasks, agents, workflows, and system operations.
    The system supports multiple execution modes, real-time communication via WebSockets,
    and comprehensive monitoring and health checking capabilities.

    ## Features
    - Task management with priority and type classification
    - Multi-agent coordination and execution
    - Real-time status updates via WebSocket
    - Comprehensive system monitoring and health checks
    - LLM configuration management
    - Batch operations and statistics
    - Meta-agent conversation management

    ## Authentication
    The API uses JWT-based authentication for securing endpoints.

    ## Rate Limiting
    Rate limiting is applied to prevent abuse and ensure system stability.

    ## WebSocket Support
    Real-time communication is supported through WebSocket endpoints for live updates.
  version: 1.0.0
  contact:
    name: LangGraph Multi-Agent Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
  /api/v1/tasks:
    get:
      tags:
        - Tasks
      summary: List tasks
      description: Retrieve a paginated list of tasks with optional filtering and sorting
      operationId: listTasks
      parameters:
        - name: status
          in: query
          description: Filter by task status
          schema:
            type: string
        - name: task_type
          in: query
          description: Filter by task type
          schema:
            type: string
            enum: [general, analysis, research, development, coordination, reporting]
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            type: integer
            minimum: 1
            maximum: 4
        - name: requester_id
          in: query
          description: Filter by requester ID
          schema:
            type: string
        - name: created_after
          in: query
          description: Filter tasks created after this timestamp
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          description: Filter tasks created before this timestamp
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            default: created_at
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            default: desc
            enum: [asc, desc]
      responses:
        '200':
          description: Successful retrieval of task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: Create a new task with specified parameters and requirements
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/agents:
    get:
      tags:
        - Agents
      summary: List agents
      description: Retrieve a list of all available agents
      operationId: listAgents
      parameters:
        - name: agent_type
          in: query
          description: Filter by agent type
          schema:
            type: string
        - name: status
          in: query
          description: Filter by agent status
          schema:
            type: string
        - name: capability
          in: query
          description: Filter by capability
          schema:
            type: string
        - name: task_id
          in: query
          description: Filter by current task
          schema:
            type: string
      responses:
        '200':
          description: Agents retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentInfo'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/llm/configs:
    get:
      tags:
        - LLM Configuration
      summary: List LLM configurations
      description: Retrieve a list of all LLM configurations
      operationId: listLLMConfigs
      responses:
        '200':
          description: LLM configurations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  configs:
                    type: array
                    items:
                      type: object
                      properties:
                        config_id:
                          type: string
                        name:
                          type: string
                        provider:
                          type: string
                        model:
                          type: string
                        active:
                          type: boolean
                        created_at:
                          type: string
                          format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - LLM Configuration
      summary: Create LLM configuration
      description: Create a new LLM configuration
      operationId: createLLMConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - provider
                - model
                - api_key
              properties:
                name:
                  type: string
                  description: Configuration name
                provider:
                  type: string
                  description: LLM provider
                  enum: [openai, anthropic, huggingface, local]
                model:
                  type: string
                  description: Model name
                api_key:
                  type: string
                  description: API key
                api_base:
                  type: string
                  description: API base URL
                parameters:
                  type: object
                  description: Model parameters
                description:
                  type: string
                  description: Configuration description
      responses:
        '201':
          description: LLM configuration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/ws:
    get:
      tags:
        - WebSocket
      summary: WebSocket endpoint
      description: Main WebSocket endpoint for real-time communication
      operationId: webSocketConnection
      parameters:
        - name: token
          in: query
          required: true
          description: Authentication token
          schema:
            type: string
      responses:
        '101':
          description: WebSocket connection established
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/mvp2/ws:
    get:
      tags:
        - MVP2 Frontend
      summary: MVP2 WebSocket endpoint
      description: WebSocket endpoint specifically for MVP2 frontend integration
      operationId: mvp2WebSocketConnection
      parameters:
        - name: token
          in: query
          required: true
          description: Authentication token
          schema:
            type: string
        - name: session_id
          in: query
          description: Session identifier
          schema:
            type: string
      responses:
        '101':
          description: WebSocket connection established
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  schemas:
    TaskCreateRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
        description:
          type: string
          minLength: 1
          maxLength: 2000
          description: Task description
        task_type:
          $ref: '#/components/schemas/TaskType'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        requirements:
          type: array
          items:
            type: string
          description: Task requirements
          maxItems: 20
        constraints:
          type: array
          items:
            type: string
          description: Task constraints
          maxItems: 10
        input_data:
          type: object
          additionalProperties: true
          description: Input data for the task
        execution_mode:
          $ref: '#/components/schemas/WorkflowExecutionMode'
        timeout_seconds:
          type: integer
          minimum: 60
          maximum: 86400
          default: 3600
          description: Timeout in seconds
        requester_id:
          type: string
          description: Requester ID
...
    TaskPriority:
      type: integer
      enum: [1, 2, 3, 4]
      description: Task priority levels
      x-enum-varnames: [LOW, MEDIUM, HIGH, URGENT]
    TaskType:
      type: string
      enum: [general, analysis, research, development, coordination, reporting]
      description: Task types
    WorkflowExecutionMode:
      type: string
      enum: [sequential, parallel, adaptive]
      description: Workflow execution modes
    TaskInfo:
      type: object
      required:
        - task_id
        - title
        - description
        - task_type
        - priority
        - status
        - created_at
        - updated_at
      properties:
        task_id:
          type: string
          description: Task ID
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Task description
        task_type:
          type: string
          description: Task type
        priority:
          type: integer
          description: Task priority
        status:
          type: string
          description: Task status
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        started_at:
          type: string
          format: date-time
          nullable: true
          description: Start timestamp
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Completion timestamp
        requester_id:
          type: string
          nullable: true
          description: Requester ID
...
    AgentInfo:
      type: object
      required:
        - agent_id
        - agent_type
        - name
        - description
        - capabilities
        - status
        - execution_stats
      properties:
        agent_id:
          type: string
          description: Agent ID
        agent_type:
          type: string
          description: Agent type
        name:
          type: string
          description: Agent name
        description:
          type: string
          description: Agent description
        capabilities:
          type: array
          items:
            type: string
          description: Agent capabilities
        status:
          type: string
          description: Agent status
        current_task:
          type: string
          nullable: true
          description: Current task
        execution_stats:
          type: object
          additionalProperties: true
          description: Execution statistics
...
    TaskListResponse:
      type: object
      required:
        - tasks
        - total
        - page
        - page_size
        - has_next
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskInfo'
          description: List of tasks
        total:
          type: integer
          description: Total number of tasks
        page:
          type: integer
          description: Current page number
        page_size:
          type: integer
          description: Number of items per page
        has_next:
          type: boolean
          description: Whether there are more pages
    ApiResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Operation success status
        message:
          type: string
          description: Response message
        data:
          type: object
          additionalProperties: true
          nullable: true
          description: Response data
        error_code:
          type: string
          nullable: true
          description: Error code if applicable
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      type: object
      required:
        - success
        - error_code
        - error_message
      properties:
        success:
          type: boolean
          default: false
          description: Always false for error responses
        error_code:
          type: string
          description: Machine-readable error code
        error_message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: BAD_REQUEST
            error_message: Invalid request parameters
            timestamp: "2024-01-01T00:00:00Z"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: UNAUTHORIZED
            error_message: Authentication required
            timestamp: "2024-01-01T00:00:00Z"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: INTERNAL_ERROR
            error_message: Internal server error
            timestamp: "2024-01-01T00:00:00Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
security:
  - BearerAuth: []
...